Reorganize messaging unit test files
